organized index.js to users.js and utils.js and login with google working with frontend
