+++ backend/api/users.js
@@ -1,175 +1,121 @@
-const MSG = require('../src/messageConstants');
+const { OAuth2Client } = require('google-auth-library');
 const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
 const speakeasy = require('speakeasy');
-
-let currentLoggedInUser = null; // Optional: to keep track of the currently logged-in user
-
-function isValidUsername(username) {
-    if (!username || username.length < 3 || username.length > 20) {
-        return false;
-    }
-    return isAlphaNumeric(username);
-}
-
-function isValidPassword(password) {
-    if (!password || password.length <= 6) {
-        return false;
-    }
-    return isSafePassword(password);
-}
-
-function isValidEmail(email) {
-    return typeof email === 'string' && /^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(email);
-}
-
-
-function isAlphaNumeric(str) {
-    return str.split('').every(char => {
-        const isLetter = (char >= 'a' && char <= 'z') || (char >= 'A' && char <= 'Z');
-        const isNumber = char >= '0' && char <= '9';
-        return isLetter || isNumber;
-    });
-}
-
-function isSafePassword(str) {
-    return str.split('').every(char => {
-        const isLetter = (char >= 'a' && char <= 'z') || (char >= 'A' && char <= 'Z');
-        const isNumber = char >= '0' && char <= '9';
-        const hasSymbols = char >= '!' && char <= '/';
-        return isLetter || isNumber || hasSymbols;
-    });
-}
-
-function generateAccessToken(user) {
-    return jwt.sign(user, process.env.ACCESS_TOKEN_SECRET, { expiresIn: '15m' });
-}
-
-function generateRefreshToken(user) {
-    return jwt.sign(user, process.env.REFRESH_TOKEN_SECRET, { expiresIn: '7d' });
-}
+const qrcode = require('qrcode');
+const MSG = require('../src/messageConstants');
+const {
+    generateAccessToken,
+    generateRefreshToken,
+    saveRefreshToken,
+    findRefreshToken,
+    deleteRefreshToken,
+    deleteAllUserRefreshTokens,
+    isValidUsername,
+    isValidPassword,
+    isValidEmail,
+    dbGet,
+    dbRun,
+    dbAll,
+    authenticateToken
+} = require('./utils');
 
 async function userRoutes(fastify, options) {
+    const db = fastify.db;
+    
+    // Google OAuth client setup
+    const client = new OAuth2Client(
+        process.env.GOOGLE_CLIENT_ID,
+        process.env.GOOGLE_CLIENT_SECRET,
+        process.env.GOOGLE_REDIRECT_URI
+    );
 
-	const db = fastify.db;
+    let currentLoggedInUser = null;
 
-    function saveRefreshToken(token, userId) {
-        return new Promise((resolve, reject) => {
-            const expiresAt = new Date();
-            expiresAt.setDate(expiresAt.getDate() + 7);
-            db.run(`INSERT INTO refresh_tokens (token, userId, expires_at) VALUES (?, ?, ?)`, 
-                [token, userId, expiresAt.toISOString()], function(err) {
-                if (err) reject(err);
-                else resolve(this.lastID);
-            });
+    // Google OAuth routes
+    fastify.get('/auth/google', (req, res) => {
+        const url = client.generateAuthUrl({
+            access_type: 'offline',
+            scope: ['profile', 'email'],
         });
-    }
-
-	fastify.get('/users/stats', (request, reply) => {
-		db.all(`SELECT * FROM userStats`, (err, rows) => {
-
-			if (err) {
-				return reply.status(500).send({ error: 'Error fetching users from database' });
-			}
-			reply.send(rows);
-		});
-	});
-
-	fastify.get('/users/stats:userId', (request, reply) => {
-		const { userId } = request.params;
-
-		db.get(`SELECT * FROM userStats WHERE userId = ?`, matchId, (err, rows) => {
-			if (err) {
-				return reply.status(500).send({ error: 'Error fetching user from database' });
-			}
-			reply.send(rows);
-		});
-	});
-
-	fastify.post('/users/stats:userId', async (request, reply) => {
-		const { userId } = request.params;
-		
-		await db.run(`INSERT INTO userStatus (userId) VALUES (?)`, [userId], function (err) {
-			if (err) {
-				return reply.status(500).send({ error: 'Error adding match to database' });
-			}
-		});
-	});
-
-	fastify.post('/users/addfriend/:userId', async (request, reply) => {
-		const { userId } = request.params;
-		
-		await db.run(`INSERT INTO userStatus (userId) VALUES (?)`, [userId], function (err) {
-			if (err) {
-				return reply.status(500).send({ error: 'Error adding match to database' });
-			}
-		});
-	});
-
-    fastify.post('/login', (req, res) => {
-        const { email, password, token } = req.body;
-    
-        if (!email || !password) {
-            return res.status(400).send({ error: 'Email and password are required' });
-        }
-    
-        if (!isValidEmail(email)) {
-            return res.status(400).send({ error: 'Invalid email format' });
-        }
-    
-        db.get(`SELECT * FROM users WHERE email = ?`, [email], async (err, row) => {
-            if (err) {
-                return res.status(500).send({ error: MSG.ERROR_FETCHING_USER });
-            }
-            if (!row) {
-                return res.status(404).send({ error: MSG.INVALID_CREDENTIALS });
-            }
-            bcrypt.compare(password, row.password, async (bcryptErr, passwordMatch) => {
-                if (bcryptErr) {
-                    return res.status(500).send({ error: MSG.ERROR_VERIFYING_PASSWORD });
-                }
-                if (!passwordMatch) {
-                    return res.status(404).send({ error: MSG.INVALID_CREDENTIALS });
-                }
-                if (row.secret) {
-                    if (!token) {
-                        return res.status(400).send({ error: MSG.AUTHENTICATOR_TOKEN_REQUIRED, requiresToken: true });
-                    }
-                    const verified = speakeasy.totp.verify({
-                        secret: row.secret,
-                        encoding: 'base32',
-                        token: token
-                    });
-                    if (!verified) {
-                        return res.status(403).send({ error: MSG.INVALID_AUTHENTICATOR_TOKEN, requiresToken: true });
-                    }
-                }
-                currentLoggedInUser = row.username;
-                try {
-                    const user = { username: row.username, userId: row.userId, email: row.email };
-                    const accessToken = generateAccessToken(user);
-                    const refreshToken = generateRefreshToken(user);
-                    await saveRefreshToken(refreshToken, row.userId);
-                    res.send({
-                        message: 'Login successful',
-                        username: row.username,
-                        accessToken: accessToken,
-                        refreshToken: refreshToken
-                    });
-                } catch (tokenError) {
-                    console.error('Error saving refresh token:', tokenError);
-                    res.status(500).send({ error: MSG.ERROR_CREATING_SESSION });
-                }
-            });
-        });
+        res.redirect(url);
     });
 
-    // --- SIGNUP ---
+    fastify.get('/auth/google/callback', async (req, res) => {
+        const code = req.query.code;
+        if (!code) {
+            return res.status(400).send(MSG.INVALID_REQUEST_NO_AUTH_CODE);
+        }
+
+        try {
+            const { tokens } = await client.getToken(code);
+            const ticket = await client.verifyIdToken({
+                idToken: tokens.id_token,
+                audience: process.env.GOOGLE_CLIENT_ID,
+            });
+
+            const payload = ticket.getPayload();
+            const googleId = payload.sub;
+            const email = payload.email;
+            const name = payload.name;
+            let user = await dbGet(db, `SELECT * FROM users WHERE googleID = ?`, [googleId]);
+            let username, userId;
+
+            if (!user) {
+                userId = await dbRun(db,
+                    `INSERT INTO users (username, password, secret, googleID, email) VALUES (?, ?, ?, ?, ?)`,
+                    [name, '', '', googleId, email]
+                );
+                username = name;
+            } else {
+                username = user.username;
+                userId = user.userId;
+            }
+
+            const accessToken = generateAccessToken({ username, userId, email });
+            const refreshToken = generateRefreshToken({ username, userId, email });
+            await saveRefreshToken(db, refreshToken, userId);
+
+            const redirectUrl = `http://localhost/?accessToken=${encodeURIComponent(accessToken)}&refreshToken=${encodeURIComponent(refreshToken)}&username=${encodeURIComponent(username)}`;
+            return res.redirect(redirectUrl);
+
+        } catch (error) {
+            console.error('Error during Google OAuth callback:', error);
+            if (error.message && error.message.includes('invalid_grant')) {
+                return res.status(400).send(MSG.AUTHORIZATION_CODE_EXPIRED);
+            }
+        }
+    });
+
+    fastify.post('/token', async (req, res) => {
+        const refreshToken = req.body.token;
+        if (!refreshToken) {
+            return res.status(401).send({ error: MSG.REFRESH_TOKEN_REQUIRED });
+        }
+
+        try {
+            const tokenData = await findRefreshToken(db, refreshToken);
+            if (!tokenData) {
+                return res.status(403).send({ error: MSG.INVALID_OR_EXPIRED_REFRESH_TOKEN });
+            }
+
+            const accessToken = generateAccessToken({ 
+                username: tokenData.username, 
+                userId: tokenData.userId,
+                email: tokenData.email
+            });
+            res.send({ accessToken: accessToken });
+        } catch (error) {
+            console.error('Error verifying refresh token:', error);
+            res.status(500).send({ error: MSG.INTERNAL_SERVER_ERROR });
+        }
+    });
+
+    // Signup
     fastify.post('/signup', (req, res) => {
         const { username, email, password } = req.body;
 
         if (!username || !email || !password) {
-            return res.status(400).send({ error: MSG.EMAIL_AND_PASSWORD_REQUIRED });
+            return res.status(400).send({ error: MSG.USERNAME_EMAIL_PASSWORD_REQUIRED });
         }
 
         if (!isValidUsername(username)) {
@@ -181,9 +127,7 @@ async function userRoutes(fastify, options) {
         }
 
         if (!isValidPassword(password)) {
-            return res.status(400).send({
-                error: MSG.PASSWORD_TOO_SHORT
-            });
+            return res.status(400).send({ error: MSG.PASSWORD_TOO_SHORT });
         }
         
         // Check for existing username or email
@@ -219,13 +163,308 @@ async function userRoutes(fastify, options) {
                                 }
                                 return res.status(500).send({ error: MSG.ERROR_SAVING_USER });
                             }
-                            res.send({ message: 'Signup successful' });
+                            res.send({ message: MSG.SIGNUP_SUCCESSFUL });
                         }
                     );
                 });
             }
         );
     });
-};
+
+    // Login
+    fastify.post('/login', (req, res) => {
+        const { email, password, token } = req.body;
+
+        if (!email || !password) {
+            return res.status(400).send({ error: MSG.EMAIL_AND_PASSWORD_REQUIRED });
+        }
+
+        if (!isValidEmail(email)) {
+            return res.status(400).send({ error: MSG.INVALID_EMAIL_FORMAT });
+        }
+
+        db.get(`SELECT * FROM users WHERE email = ?`, [email], async (err, row) => {
+            if (err) {
+                return res.status(500).send({ error: MSG.ERROR_FETCHING_USER });
+            }
+            if (!row) {
+                return res.status(404).send({ error: MSG.INVALID_CREDENTIALS });
+            }
+            bcrypt.compare(password, row.password, async (bcryptErr, passwordMatch) => {
+                if (bcryptErr) {
+                    return res.status(500).send({ error: MSG.ERROR_VERIFYING_PASSWORD });
+                }
+                if (!passwordMatch) {
+                    return res.status(404).send({ error: MSG.INVALID_CREDENTIALS });
+                }
+                if (row.secret) {
+                    if (!token) {
+                        return res.status(400).send({ error: MSG.AUTHENTICATOR_TOKEN_REQUIRED, requiresToken: true });
+                    }
+                    const verified = speakeasy.totp.verify({
+                        secret: row.secret,
+                        encoding: 'base32',
+                        token: token
+                    });
+                    if (!verified) {
+                        return res.status(403).send({ error: MSG.INVALID_AUTHENTICATOR_TOKEN, requiresToken: true });
+                    }
+                }
+                currentLoggedInUser = row.username;
+                try {
+                    const user = { username: row.username, userId: row.userId, email: row.email };
+                    const accessToken = generateAccessToken(user);
+                    const refreshToken = generateRefreshToken(user);
+                    await saveRefreshToken(db, refreshToken, row.userId);
+                    res.send({
+                        message: MSG.LOGIN_SUCCESSFUL,
+                        username: row.username,
+                        accessToken: accessToken,
+                        refreshToken: refreshToken
+                    });
+                } catch (tokenError) {
+                    console.error('Error saving refresh token:', tokenError);
+                    res.status(500).send({ error: MSG.ERROR_CREATING_SESSION });
+                }
+            });
+        });
+    });
+
+    // Logout
+    fastify.post('/logout', async (req, res) => {
+        const { token } = req.body;
+        currentLoggedInUser = null;
+
+        if (token) {
+            try {
+                await deleteRefreshToken(db, token);
+            } catch (error) {
+                console.error('Error deleting refresh token:', error);
+            }
+        }
+
+        res.send({ message: MSG.LOGOUT_SUCCESSFUL });
+    });
+
+    fastify.get('/current-user', { preHandler: authenticateToken }, (req, res) => {
+        res.send({ username: req.user.username, userId: req.user.userId, email: req.user.email });
+    });
+
+    fastify.get('/users', { preHandler: authenticateToken }, async (req, res) => {
+        try {
+            const users = await dbAll(db, `SELECT userId, username, secret, googleID, email FROM users`);
+            res.send(users);
+        } catch (error) {
+            console.error('Database error in /users endpoint:', error);
+            res.status(500).send({ error: MSG.ERROR_FETCHING_USERS });
+        }
+    });
+
+    // 2FA routes
+    fastify.post('/generate', { preHandler: authenticateToken }, (request, reply) => {
+        const { email } = request.body;
+
+        if (!email) {
+            return reply.status(400).send({ error: MSG.EMAIL_REQUIRED });
+        }
+
+        const userEmail = email || request.user.email;
+
+        if (request.user.email !== userEmail) {
+            return reply.status(403).send({ error: MSG.CAN_ONLY_GENERATE_FOR_OWN_ACCOUNT });
+        }
+
+        db.get(`SELECT * FROM users WHERE email = ?`, [userEmail], (err, row) => {
+            if (err) {
+                return reply.status(500).send({ error: MSG.ERROR_FETCHING_USER });
+            }
+            if (!row) {
+                return reply.status(404).send({ error: MSG.USER_NOT_FOUND });
+            }
+
+            const secret = speakeasy.generateSecret({ name: "ft_transcendence(" + row.username + ")" });
+
+            qrcode.toDataURL(secret.otpauth_url, function (err, qrCodeUrl) {
+                if (err) {
+                    return reply.status(500).send({ error: MSG.ERROR_GENERATING_QR });
+                }
+                reply.send({ qrCodeUrl, secret: secret.base32 });
+            });
+        });
+    });
+
+    fastify.post('/verify-token', (req, res) => {
+        const { email, token, secret } = req.body;
+
+        if (!email || !token) {
+            return res.status(400).json({ error: MSG.EMAIL_AND_TOKEN_REQUIRED });
+        }
+
+        if (secret) {
+            const verified = speakeasy.totp.verify({
+                secret: secret,
+                encoding: 'base32',
+                token: token,
+                window: 2
+            });
+
+            if (verified) {
+                db.run(`UPDATE users SET secret = ? WHERE email = ?`, [secret, email], function (err) {
+                    if (err) {
+                        return res.status(500).json({ error: MSG.ERROR_SAVING_SECRET });
+                    }
+                    res.json({ message: MSG.TOKEN_VERIFIED_2FA_ENABLED });
+                });
+            } else {
+                res.status(403).json({ error: MSG.INVALID_TOKEN });
+            }
+        } else {
+            db.get(`SELECT secret FROM users WHERE email = ?`, [email], (err, row) => {
+                if (err) {
+                    return res.status(500).json({ error: MSG.ERROR_FETCHING_SECRET });
+                }
+                if (!row || !row.secret) {
+                    return res.status(404).json({ error: MSG.USER_NOT_FOUND_OR_NO_2FA });
+                }
+                const verified = speakeasy.totp.verify({
+                    secret: row.secret,
+                    encoding: 'base32',
+                    token: token
+                });
+                if (verified) {
+                    res.json({ message: MSG.TOKEN_VERIFIED_SUCCESSFULLY });
+                } else {
+                    res.status(403).json({ error: MSG.INVALID_TOKEN });
+                }
+            });
+        }
+    });
+
+    fastify.get('/check-2fa', (req, res) => {
+        const { email } = req.query;
+
+        if (!email) {
+            return res.status(400).send({ error: MSG.EMAIL_REQUIRED });
+        }
+
+        db.get(`SELECT secret FROM users WHERE email = ?`, [email], (err, row) => {
+            if (err) {
+                return res.status(500).send({ error: MSG.ERROR_FETCHING_USER });
+            }
+            if (!row) {
+                return res.status(404).send({ error: MSG.USER_NOT_FOUND });
+            }
+            res.send({ has2FA: !!row.secret });
+        });
+    });
+
+    fastify.get('/current-token', { preHandler: authenticateToken }, (req, res) => {
+        const { email } = req.query;
+
+        if (!email) {
+            return res.status(400).send({ error: MSG.EMAIL_REQUIRED });
+        }
+
+        if (req.user.email !== email) {
+            return res.status(403).send({ error: MSG.CAN_ONLY_GET_TOKENS_FOR_OWN_ACCOUNT });
+        }
+
+        db.get(`SELECT secret FROM users WHERE email = ?`, [email], (err, row) => {
+            if (err) {
+                return res.status(500).send({ error: MSG.ERROR_FETCHING_SECRET });
+            }
+            if (!row || !row.secret) {
+                return res.status(404).send({ error: MSG.USER_NOT_FOUND_OR_NO_2FA });
+            }
+            const token = speakeasy.totp({
+                secret: row.secret,
+                encoding: 'base32'
+            });
+            res.send({ token });
+        });
+    });
+
+    fastify.post('/change-username', { preHandler: authenticateToken }, (request, reply) => {
+        const { newUsername } = request.body;
+        if (!newUsername || !isValidUsername(newUsername)) {
+            return reply.status(400).send({ error: MSG.INVALID_NEW_USERNAME });
+        }
+        db.run(
+            `UPDATE users SET username = ? WHERE userId = ?`,
+            [newUsername, request.user.userId],
+            function (err) {
+                if (err) {
+                    if (err.code === 'SQLITE_CONSTRAINT') {
+                        return reply.status(400).send({ error: MSG.USERNAME_EXISTS });
+                    }
+                    return reply.status(500).send({ error: MSG.DATABASE_ERROR });
+                }
+                reply.send({ message: MSG.USERNAME_UPDATED_SUCCESSFULLY });
+            }
+        );
+    });
+
+    fastify.post('/change-password', { preHandler: authenticateToken }, (request, reply) => {
+        const { newPassword } = request.body;
+        if (!newPassword || !isValidPassword(newPassword)) {
+            return reply.status(400).send({ error: MSG.INVALID_NEW_PASSWORD });
+        }
+        bcrypt.hash(newPassword, 13, (err, hashedPassword) => {
+            if (err) {
+                return reply.status(500).send({ error: MSG.ERROR_HASHING_PASSWORD });
+            }
+            db.run(
+                `UPDATE users SET password = ? WHERE userId = ?`,
+                [hashedPassword, request.user.userId],
+                function (err) {
+                    if (err) {
+                        reply.status(500).send({ error: MSG.DATABASE_ERROR });
+                        return;
+                    }
+                    reply.send({ message: MSG.PASSWORD_UPDATED_SUCCESSFULLY });
+                }
+            );
+        });
+    });
+
+    // User stats routes
+    fastify.get('/users/stats', (request, reply) => {
+        db.all(`SELECT * FROM userStats`, (err, rows) => {
+            if (err) {
+                return reply.status(500).send({ error: MSG.ERROR_FETCHING_USERS_FROM_DATABASE });
+            }
+            reply.send(rows);
+        });
+    });
+
+    fastify.get('/users/stats/:userId', (request, reply) => {
+        const { userId } = request.params;
+
+        db.get(`SELECT * FROM userStats WHERE userId = ?`, [userId], (err, row) => {
+            if (err) {
+                return reply.status(500).send({ error: MSG.ERROR_FETCHING_USER_FROM_DATABASE });
+            }
+            reply.send(row);
+        });
+    });
+
+    fastify.post('/users/stats/:userId', async (request, reply) => {
+        const { userId } = request.params;
+        
+        db.run(`INSERT INTO userStats (userId) VALUES (?)`, [userId], function (err) {
+            if (err) {
+                return reply.status(500).send({ error: MSG.ERROR_ADDING_USER_STATS_TO_DATABASE });
+            }
+            reply.send({ message: MSG.USER_STATS_CREATED_SUCCESSFULLY });
+        });
+    });
+
+    fastify.post('/users/addfriend/:userId', async (request, reply) => {
+        const { userId } = request.params;
+        
+//missing code
+        reply.send({ message: MSG.FRIEND_FUNCTIONALITY_NOT_IMPLEMENTED });
+    });
+}
 
 module.exports = userRoutes;
