+++ frontend/src/main.ts
@@ -36,7 +36,6 @@ const routes = {
 type NavigateFn = (path: string) => void;
 export let navigate: NavigateFn;
 
-// listen for notifications and show them in UI
 notificationService.listen((notif) => {
   window.dispatchEvent(new CustomEvent("new-notification", { detail: notif }));
 });
@@ -45,6 +44,56 @@ notificationService.listen((notif) => {
 startMockNotifications();
 
 document.addEventListener('DOMContentLoaded', async () => {
+  const params = new URLSearchParams(window.location.search);
+  const accessToken = params.get('accessToken');
+  const refreshToken = params.get('refreshToken');
+  const username = params.get('username');
+  
+  if (accessToken && refreshToken && username) {
+    console.log('Google OAuth tokens detected, processing login...');
+    
+    const extractEmailFromToken = (token: string): string => {
+      try {
+        const payload = JSON.parse(atob(token.split('.')[1]));
+        return payload.email || '';
+      } catch (error) {
+        console.error('Error decoding token:', error);
+        return '';
+      }
+    };
+    
+    const email = extractEmailFromToken(accessToken);
+    
+    // Store tokens in localStorage
+    localStorage.setItem('accessToken', accessToken);
+    localStorage.setItem('refreshToken', refreshToken);
+    localStorage.setItem('loggedInUser', decodeURIComponent(username));
+    localStorage.setItem('userEmail', email);
+    
+    state.userData = {
+      username: decodeURIComponent(username),
+      email: email,
+      accessToken: accessToken,
+      refreshToken: refreshToken,
+    };
+    
+    // Clean URL (remove tokens from address bar)
+    window.history.replaceState({}, document.title, "/");
+    
+    window.dispatchEvent(new CustomEvent('login-success', { 
+      bubbles: true, 
+      composed: true,
+      detail: { 
+        username: decodeURIComponent(username), 
+        email,
+        accessToken, 
+        refreshToken 
+      } 
+    }));
+    
+    console.log('Google OAuth login successful for:', decodeURIComponent(username));
+  }
+
   // translations by default will be in english
   // TODO: populate state, watch for changes? (?)
   state.translations = await getTranslations(state.language);
@@ -55,10 +104,28 @@ document.addEventListener('DOMContentLoaded', async () => {
   navigate = initRouter(routes, 'app');
 
   window.addEventListener('login-success', () => {
-    renderLoading('app'); // render loading state, transition between pages
+    renderLoading('app'); 
     setTimeout(() => {
       navigate('/');
     }, 1200);
+  });
+
+  window.addEventListener('logout', () => {
+    console.log('Logout event triggered, clearing user data...');
+    
+    localStorage.removeItem('accessToken');
+    localStorage.removeItem('refreshToken');
+    localStorage.removeItem('loggedInUser');
+    localStorage.removeItem('userEmail');
+    
+    state.userData = null;
+    
+    renderLoading('app');
+    setTimeout(() => {
+      navigate('/login');
+    }, 1200);
+    
+    console.log('User logged out successfully');
   });
 
   window.addEventListener('languagechange', async (e: Event) => {
