+++ frontend/src/pages/home.ts
@@ -2,6 +2,7 @@ import '@/components/ThemeToggle.js';
 import '@/components/LanguagesDropdown.js';
 import '@/components/TopBar.js';
 import '@/components/NotificationsBar.js';
+import { logout } from '@/services/authService.js';
 
 export function renderHome(containerId: string) {
   const container = document.getElementById(containerId);
@@ -20,7 +21,7 @@ export function renderHome(containerId: string) {
         <theme-toggle slot="toggle"></theme-toggle>
         <languages-dropdown slot="language"></languages-dropdown>
         <img slot="avatar" src="https://api.dicebear.com/7.x/pixel-art/svg?seed=robot" alt="Avatar" />
-        <button slot="logout" onclick="/* your logout logic */">
+        <button slot="logout" id="logoutBtn">
           <img src="https://unpkg.com/pixelarticons@1.8.1/svg/logout.svg" alt="Logout" style="width:22px;height:22px;filter:invert(1);" />
           Logout
         </button>
@@ -32,6 +33,18 @@ export function renderHome(containerId: string) {
         <div class="intro-text text-4xl text-white text-center z-10">[PLACEHOLDER]</div>
       </section>
     `;
+
+    const logoutBtn = container.querySelector('#logoutBtn');
+    if (logoutBtn) {
+      logoutBtn.addEventListener('click', async () => {
+        const refreshToken = localStorage.getItem('refreshToken');
+        if (refreshToken) {
+          await logout(refreshToken);
+        } else {
+          window.dispatchEvent(new CustomEvent('logout', { bubbles: true, composed: true }));
+        }
+      });
+    }
   }
 }
