+++ frontend/src/services/authService.ts
@@ -6,12 +6,11 @@ export interface AuthResponse {
   requiresToken?: boolean;
   has2FA?: boolean;
   qrCodeUrl?: string;
+  secret?: string;
   accessToken?: string;
   refreshToken?: string;
 }
 
-// const BASE_URL = "http://localhost:4000";
-// const BASE_URL = "http://192.168.64.26:4000";
 const BASE_URL = "/api";
 
 export async function signup(username: string, email: string, password: string) {
@@ -42,25 +41,29 @@ export async function login(email: string, password: string, token?: string): Pr
 }
 
 export async function googleLogin(): Promise<void> {
-  window.location.href = `${BASE_URL}/auth/google`;
+  window.location.href = `/auth/google`;
 }
 
 export async function logout(refreshToken: string): Promise<AuthResponse> {
-  const res = await fetch(`${BASE_URL}/logout`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ token: refreshToken }),
-  });
-  return await res.json();
-}
-
-export async function checkServerLoginStatus(accessToken: string): Promise<AuthResponse> {
-  const res = await fetch(`${BASE_URL}/current-user`, {
-    headers: {
-      'Authorization': `Bearer ${accessToken}`
-    }
-  });
-  return await res.json();
+  try {
+    const res = await fetch(`${BASE_URL}/logout`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ token: refreshToken }),
+    });
+    
+    const result = await res.json();
+    
+    window.dispatchEvent(new CustomEvent('logout', { bubbles: true, composed: true }));
+    
+    return result;
+  } catch (error) {
+    console.error('Error during logout:', error);
+    
+    window.dispatchEvent(new CustomEvent('logout', { bubbles: true, composed: true }));
+    
+    return { error: 'Logout completed locally' };
+  }
 }
 
 export async function check2FAStatus(email: string): Promise<AuthResponse> {
@@ -120,4 +123,21 @@ export async function changePassword(newPassword: string, accessToken: string): 
     body: JSON.stringify({ newPassword }),
   });
   return await res.json();
+}
+
+export async function checkServerLoginStatus(): Promise<AuthResponse> {
+  const token = localStorage.getItem('accessToken');
+  try {
+    const response = await fetch(`${BASE_URL}/current-user`, {
+      headers: {
+        'Authorization': `Bearer ${token}`
+      }
+    });
+    if (response.ok) {
+      return await response.json();
+    }
+    throw new Error('Not authenticated');
+  } catch (error: any) {
+    return { error: error.message || 'Network error' };
+  }
 }
