+++ nginx.conf
@@ -3,10 +3,23 @@ events {}
 http {
     # Reduce log verbosity
     error_log /var/log/nginx/error.log warn;
-    access_log /var/log/nginx/access.log;
+    access_log off;
+    # access_log /var/log/nginx/access.log;
 
     server {
         listen 80;
+
+        location /auth/google {
+            proxy_pass http://backend:4000/api/auth/google;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+        }
+
+        location /auth/google/callback {
+            proxy_pass http://backend:4000/api/auth/google/callback;
+            proxy_set_header Host $host;
+            proxy_set_header X-Real-IP $remote_addr;
+        }
 
         location /api/ {
             proxy_pass http://backend:4000/api/;
@@ -29,4 +42,4 @@ http {
             proxy_set_header X-Real-IP $remote_addr;
         }
     }
-} 
+}
