+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "node-fetch": "^3.3.2"
+  }
+}
